refactor: extract false negative analysis and session hook
