+++ src/hooks/useSessionManager.test.js
@@ -1,0 +1,86 @@
+import { describe, it, expect, vi, beforeEach, afterEach } from 'vitest';
+import { renderHook, act } from '@testing-library/react';
+import { useSessionManager } from './useSessionManager';
+import { buildSession } from '../utils/session';
+
+const memoryStore = { last: null };
+
+vi.mock('../utils/db', () => ({
+  putLastSession: vi.fn(async (session) => {
+    memoryStore.last = session;
+  }),
+  getLastSession: vi.fn(async () => memoryStore.last),
+  clearLastSession: vi.fn(async () => {
+    memoryStore.last = null;
+  }),
+}));
+
+describe('useSessionManager', () => {
+  beforeEach(() => {
+    memoryStore.last = null;
+    try {
+      window.localStorage.removeItem('persistEnabled');
+    } catch {
+      // ignore
+    }
+    vi.useFakeTimers();
+  });
+  afterEach(() => {
+    vi.useRealTimers();
+  });
+
+  it('auto-saves when enabled', async () => {
+    const { result } = renderHook(() =>
+      useSessionManager({
+        summaryData: [],
+        detailsData: [],
+        clusterParams: {},
+        dateFilter: { start: null, end: null },
+        rangeA: { start: null, end: null },
+        rangeB: { start: null, end: null },
+        fnPreset: 'balanced',
+        setClusterParams: vi.fn(),
+        setDateFilter: vi.fn(),
+        setRangeA: vi.fn(),
+        setRangeB: vi.fn(),
+        setSummaryData: vi.fn(),
+        setDetailsData: vi.fn(),
+      })
+    );
+    act(() => result.current.setPersistEnabled(true));
+    vi.advanceTimersByTime(600);
+    const { putLastSession } = await import('../utils/db');
+    expect(putLastSession).toHaveBeenCalledTimes(1);
+  });
+
+  it('loads a saved session', async () => {
+    memoryStore.last = buildSession({
+      summaryData: [{ AHI: '1' }],
+      detailsData: [],
+    });
+    const setSummaryData = vi.fn();
+    const { result } = renderHook(() =>
+      useSessionManager({
+        summaryData: [],
+        detailsData: [],
+        clusterParams: {},
+        dateFilter: { start: null, end: null },
+        rangeA: { start: null, end: null },
+        rangeB: { start: null, end: null },
+        fnPreset: 'balanced',
+        setClusterParams: vi.fn(),
+        setDateFilter: vi.fn(),
+        setRangeA: vi.fn(),
+        setRangeB: vi.fn(),
+        setSummaryData,
+        setDetailsData: vi.fn(),
+      })
+    );
+    await act(async () => {
+      await result.current.handleLoadSaved();
+    });
+    const { getLastSession } = await import('../utils/db');
+    expect(getLastSession).toHaveBeenCalledTimes(1);
+    expect(setSummaryData).toHaveBeenCalled();
+  });
+});
